# Fixed a long running query when using find_each