## Description
Our production `push_message` table has now grown to over 6.5 million records and causing issues when using [find_each](http://api.rubyonrails.org/classes/ActiveRecord/Batches.html#method-i-find_each) method from ruby. There isn't a way of specifying the `ORDER BY` clause which causes the query to check for push messages to take an average of 6-7 seconds with 6.5> million records.

This fix takes the query times from an average of 6-7 seconds down to about an average of 6-10 milliseconds.

## Data from running these queries in Production
### Previous Query: 
```
EXPLAIN ANALYZE SELECT "push_messages".* FROM "push_messages" WHERE (delivered = 'f' AND failed = 'f' AND (deliver_after IS NULL OR deliver_after < '2017-01-26 11:20:21.180131')) ORDER BY "push_messages"."id" ASC LIMIT 1000;
                                                                           QUERY PLAN                                                                            
-----------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.43..15836.61 rows=1000 width=512) (actual time=9406.990..9406.990 rows=0 loops=1)
   ->  Index Scan using push_messages_pkey on push_messages  (cost=0.43..4467559.49 rows=282111 width=512) (actual time=9406.986..9406.986 rows=0 loops=1)
         Filter: ((NOT delivered) AND (NOT failed) AND ((deliver_after IS NULL) OR (deliver_after < '2017-01-26 11:20:21.180131'::timestamp without time zone)))
         Rows Removed by Filter: 6578264
 Planning time: 0.238 ms
 Execution time: 9407.042 ms
(6 rows)
```

### New Query:
```
EXPLAIN ANALYZE SELECT "push_messages".* FROM "push_messages" WHERE (delivered = 'f' AND failed = 'f' AND (deliver_after IS NULL OR deliver_after < '2017-01-26 11:20:21.180131'));
                                                                                          QUERY PLAN                                                                                           
-----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Limit  (cost=0.43..1788.71 rows=1000 width=512) (actual time=6.056..6.056 rows=0 loops=1)
   ->  Index Scan using index_push_messages_on_delivered_and_failed_and_deliver_after on push_messages  (cost=0.43..504492.73 rows=282111 width=512) (actual time=6.051..6.051 rows=0 loops=1)
         Index Cond: ((delivered = false) AND (failed = false))
         Filter: ((NOT delivered) AND (NOT failed) AND ((deliver_after IS NULL) OR (deliver_after < '2017-01-26 11:20:21.180131'::timestamp without time zone)))
 Planning time: 0.223 ms
 Execution time: 6.108 ms
(6 rows)
```
## Benchmark Results
```
Rehearsal ----------------------------------------------------------
push message find_each 185.510000   8.940000 194.450000 (996.563827)
----------------------------------------------- total: 194.450000sec

                             user     system      total        real
push message find_each   0.000000   0.000000   0.000000 (  0.010294)
```

```
Rehearsal ----------------------------------------------------------
push message in_groups 180.630000   8.640000 189.270000 (988.253963)
----------------------------------------------- total: 189.270000sec

                             user     system      total        real
push message in_groups   0.010000   0.000000   0.010000 (  0.010251)
```