@gwynantj 2017-01-30
@brentmurphy don't know where to start adding tests tbh. It's a library that hasn't got any specs. If I had time I would add some but the deals work has a bigger priority at the moment. And this just looked like an easy win.