@brentmurphy 2017-01-30
Fair enough and agreed on leaving it. Odd for tompesman/push-core not to have tests.